Continuing refactor, isolated rendering to separate file
